Feature: Make Logger a Separate Package
