+++ ToDo/Views/ToDoItemDetailModule/ToDoItemDetail.swift
@@ -1,5 +1,6 @@
 import SwiftUI
 import SwiftData
+import LoggerPackage
 
 struct ToDoItemDetail: View {
     @Query(sort: \Category.id) private var categories: [Category]
